Added Side menu and top nav bar for home page
The new side menu is currently only consists of the home and calendar
tabs but the buttons have not been implemented yet. I will implement
them soon after reading up more on the router.

The top nav bar is reusable and can be used for most if not all the
pages. It consists of a menu button which toggles the side menu as well
as a add button which can be used to add items to the calendar. However
the logic for adding items as well as toggling the menu bar is still
currently in home.tsx and should be transferred to app.tsx or some other
page higher up to be reused in multiple pages.