Edited home page and added daily schedule
Abstracted the Addtask modal and the upcoming deadlines components out
of the home page.

Added support for displaying daily schedule, adding items to schedule
through the addTask modal as well. The addTask modal will add task to
upcoming deadlines if the "add to schdule" checkbox is unchecked, and
will add to daily schedule otherwise.